Fix nested template genearation when attribute path is not string
When determining attributes to cache, we seem to be assuming the
paths to be strings. If there are resource references in outputs
it would be None during validation.

As we've the same code snippet in all group resources
ex. ResourceGroup, ResourceChain and heat AutoScalingGroup, this
fixes it in all those places.

Change-Id: I36d623b183d01632041113f7bff63bf255f5d53c
Closes-Bug: #1719603
(cherry picked from commit aabb3392b30e4905e8b780bec4c819ec6058b2b6)