Handle resource plugins without default_client_name
We want to be able to call is_not_found() or is_conflict() from the client
plugin from the Resource class, regardless of what resource plugin is
implementing the resource type. This is only possible when the
default_client_name is set, which is optional for plugins to do. As a
result, when handling an exception we need to check explicitly that there
is a default_client name, or (as in Resource.validate_external()) we will
get an assertion error.

To simplify the avoidance of programming errors, define a
_default_client_plugin() method that *always* returns a plugin whose
is_not_found() and is_conflict() methods can be called.

Change-Id: I1005397187457d9be42ba86096644dd17287d0e7
Story: #1745024
Task: 17677