Improve error reporting for missing nested template
When the template for a TemplateResource was missing from the files dict
and not otherwise available, we provided misleading feedback when
validating the parent template.

First, we log that we are fetching the template from a URL, even though we
would not actually attempt to do so if it is a file:/// URL. Move the log
to after the allowed scheme check, and log debug to indicate when we cannot
find a template file.

While an exception would be generated in
TemplateResource._generate_schema() at resource initialisation time, it is
suppressed and saved for later. An empty template is used to generate the
properties schema. The exception is re-raised when validate() is called,
but during a template validation we only call validate_template(), not
validate(), so the first error we run into will be a mismatch of property
names. (If no property values were passed in the parent template, we may
not even get an error even though we won't be able to create a stack from
the given data.) Also re-raise the stored exception at the beginning of
validate_template() so that users will see the true source of the error.

Change-Id: I1bc100684e1b84fc9ac54ef523d798b317e4dc51
Story: #1739447
Task: 22219