Handle data inconsistency in grouputils
In convergence, retrieving the template for a stack will return the latest
template submitted by the user and accepted to start a traversal. If this
template removes resources from the stack, they may continue to appear as
available when getting a list of current resources even if a definition
for them doesn't appear in the current template.

This occasionally caused a KeyError when trying to match up data from the
nested stack to the nested stack's template in grouputils using the
GroupInspector. Fix this by accounting for the fact that the stack and
template may contain different sets of resources.

Change-Id: I86d80ab766667194f4e6e3b6112d6b1b7a3733d8
Closes-Bug: #1747365