Always use string ID for WaitConditionHandle signals
Using an integer as the ID - which is what we currently do for
OS::Heat::WaitConditionHandles without an ID explicitly supplied by the
signaller - means that we can implicitly create an entry that will collide
with an explicitly specified string ID when we serialise to and then
deserialise from JSON to store the metadata in the database. These types of
collisions are silent and avoid being logged (despite the fact that we
attempt to catch and log collisions), which makes any issues very, very
difficult to debug.

Change-Id: Ia613e79dc9757c4594b84ed2105bfc1357d1e037
Related-Bug: #1738653