Increment resource atomic key when storing attributes
When we store resource attribute values in the database, we increment the
atomic_key in the DB. This means we should also increment the _atomic_key
member of the Resource object, so that any future writes to the database
will expect the correct key.

In practice, this doesn't matter because there's no RPC API call that
stores attributes and then writes to the resource in the database for other
reasons. (Attributes are updated if they are missing on show-stack with
outputs, which is otherwise a read-only call; they're also updated at the
very end of a check-resource call in convergence.) But by maintaining a
consistent model this could help to prevent future bugs.

Change-Id: I71ed5907d33f2293c04ad3bace639b1d114e9a77
Related-Bug: #1763021