Remove install-requirements script
this script is not used anywhere anymore, and it will not work
with recent pip 10.x release due to changes in the pip API.

Initially this script was meant to install only a limited set of requirements
to run heat_integrationtests without having to install whole Heat.

Since the split of heat_tempest_plugin the tests left in the
heat_integrationtests are meant to be run only in a DevStack-like env,
so all the requirements are already available.

Change-Id: I94c82fe100dfb6557cc960ef1198f1195780f28f