External_fixed_ips break router creation.
In the _resolve_subnet method the code iterates over a dictionary
while modifying it using the "pop" command. This is incorrect and
causes a Runtime error.

The fix is to make a copy, iterate over that  while manipulating the original
dictionary. Python 2.7 can do this using iter(), however python 3 would not
have worked. The fix does a copy first which makes the code operate the same
in both interpreters, since there is no "six" routine for this.

Change-Id: I11468bfbe7fafd5c0d79138a94520dec1aafbc04
Story: 2002549
Task: 22111