Unlimited cinder quotas throws exception
Cinder-quota-resource is supposed to support a value of -1 to indicate an
unlimited quota for a particular key (volumes, gigabytes or snapshots).
However the implementation had an iterator error (you cannot modify a
dict as you are iterating over it) and no unit tests to expose that bug.

This commit fixes the issue.

Change-Id: Iea28fdc38c51df5e87e6957a6e4797feea9057ee
Story: 2002550
Task: 22112