Fix purging events from stacks converted to convergence
Stacks with convergence enabled are guaranteed not to have a backup stack,
but they may have previously had backup stacks (prior to being converted to
convergence) that remain only soft-deleted. In this case they may still
fail the foreign key constraints in the database when purging events, so
don't bail out of the workaround path when we encounter this situation.

This will allow events to be purged from these stacks without having to
wait for purge_deleted to purge the backup stacks first.

Change-Id: If541d421a0a2af17ab937ec0a27bee6e4e9c1f6b
Story: #2002643
Task: 22648