Don't get frozen definitions for resource being deleted
At the start of a legacy stack update, we calculate the frozen definitions
for all of the existing resources. This is so that we can later compare
them against the new resources to see if they need to be updated.

Since resources that are being deleted (i.e. don't appear in the new
template) won't ever need to be compared to the new definitions, there's no
need for us to calculate this data.

Apparently there are (extremely rare) situations where getting the frozen
definitions will fail, so this at least gives users an opportunity to
recover by deleting only the resources causing trouble rather than the
whole stack.

Change-Id: I04bed83cc2181e1e61bf90c2f259cfd123362b06
Story: 2002966
Task: 22972