Don't change INIT_COMPLETE resources to CHECK_COMPLETE
We assume that only resources that are in INIT_COMPLETE need to be created
(as opposed to updated). If we run a check on a resource that hasn't been
created yet, we shouldn't change the status to CHECK_COMPLETE because
thereafter we will try to update it.

Changing the status to CHECK_FAILED (as we do if the resource supports
handle_check()) is OK because the resource will get replaced on the next
update anyway.

Change-Id: Iaadf3b57b0e8edc76d338c017dbd05bf2a37698f
Story: #2003062
Task: 23102