Always start convergence worker service
The purpose of the cfg.CONF.convergence_engine option is to determine
whether new stacks will be created using convergence or not. For some
reason, however, the worker service was not started when the option was
disabled, meaning that previously-created convergence stacks could no
longer be updated (there was a hack in place to allow them to be deleted).

Always start the worker service and continue operating any existing
convergence stacks in convergence mode, regardless of the current setting
for new stacks.

Change-Id: Ic3574e4d15ac48b2bc8e0e8101c81d24f40f0606
Related-Bug: #1508324