Delete snapshots using contemporary resources
When deleting a snapshot, we used the current resources in the stack to
call delete_snapshot() on. However, there is no guarantee that the
resources that existed at the time the snapshot was created were of the
same type as any current resources of the same name.

Use resources created using the template in the snapshot to do the
snapshot deletion.

This also solves the problem addressed in
df1708b1a83f21f249aa08827fa88a25fc62c9e5, whereby snapshots had to be
deleted before the stack deletion was started in convergence because
otherwise the 'latest' template contained no resources.

That allows us to once again move the snapshot deletion after the start of
the stack deletion, which is consistent with when it happens in the legacy
path. Amongst other things, this ensures that any failures can be reported
correctly.

Change-Id: I1d239e9fcda30fec4795a82eba20c3fb11e9e72a