Refactor deferral of stack state persistence
When we hold a StackLock, we defer any persistence of COMPLETE or FAILED
states in state_set() until we release the lock, to avoid a race on the
client side. The logic for doing this was scattered about and needed to be
updated together, which has caused bugs in the past.

Collect all of the logic into a single implementation, for better
documentation and so that nothing can fall through the cracks.

Change-Id: I6757d911a63708a6c6356f70c24ccf1d1b5ec076