Workaround client race in legacy nested stack delete
When we delete a legacy stack, the delete call may return when a previous
delete is still IN_PROGRESS and we are waiting for it to be cancelled - for
up to 250s by default plus the time it takes to kill any remaining threads.
This means that if the client subsequently sees a DELETE_FAILED status, it
can be ambiguous as to whether it refers to the previous delete or the one
it just commanded.

Work around this when Heat itself is the client (i.e. we are deleting a
nested stack).

First, put the time the delete started into the status_reason message when
the stack goes into the DELETE_IN_PROGRESS state, so that we can
distinguish.

Then, require the stack to report its status as DELETE_FAILED twice in a
row, 10s apart, before we mark it as failed - unless we have definitively
seen a different delete commence in the meantime.

Change-Id: Ie23ba5f4538a8f5c2e7d64caa6cbe530c013b520
Story: #1669608
Task: 17214