Unit tests: Fix mock errors with too few side effects
When a mock hasn't been provided with a long-enough list of side-effects
for the number of times it is called, it raises StopIteration (instead
of something sensible like AssertionError). Prior to Python 3.7, this
exception could bubble up until it just stopped a task. (In Python 3.7
it will eventually be caught and turned into a RuntimeError.)

To ensure that any errors of this sort are not handled silently, and to
enable us to test for them prior to using Python 3.7, catch any
StopIteration errors coming from plugin-provided non-generator functions
and convert them to RuntimeError exceptions.

This reveals many errors in the unit tests, which are also fixed by this
patch. This backport contains fewer fixes than master, because many of
the errors were introduced in the process of moving from mox to mock in
the Rocky release cycle.

Change-Id: I5a1eff6b704dff7c17edcbbe58cdbc380ae6abc9
Story: #2003412
Task: 24553
(cherry picked from commit 38fad07c0a01072ee4e17815f6ad0e5c2ce1f64a)