Fail to release a stack_lock in the database is a fatal action.
When a heat-engine thread activity completes, it calls release on
its stack_lock object in the database. If that release action fails
due to an inability to update the database, that engine process is no
longer usable. This code catches that failure, logs it, and terminates
that engine process so that a new one can be started. New heat engines
will automatically purge stale stack_locks from the database.
Also, make sure that if the thread exit does not teardown the process
after 5 seconds, the non blockable os level exit call will be invoked.

This bug is very timing specific. The DB error needs to exist when the
stack_lock release fails

Change-Id: I7663b2270bf325cd8e3dd194f2994227fd6f5e8a
Story: 2003439
Task: 24635