Update resource definitions after legacy in-place update
Since we remove a resource's definition after it is deleted, we must also
update a resource's definition after it is updated. Otherwise, a resource
that depended on the now-deleted resource prior to its last in-place update
could remain in the template, leaving the stack inoperable since it is
unable to calculate its own dependencies.

We already update the template in the backup stack when updating resources
in-place, and in the main stack when creating replacement resources. Also
update the main template when resources are updated in-place, not just when
they are replaced.

Change-Id: I97ddc6a970ffd27bfd40bb8f6bec8d1cbf28a815
Story: #2003612
Task: 24946