Calculate resource data prior to legacy updates
Ensure that the StackDefinition is populated with all of the resource data
(i.e. refid and attribute values) prior to doing a legacy update. This
ensures that it is available for the purposes of obtaining the frozen
definitions of the existing resources prior to starting the update.

Obtaining the frozen definition may rely on resolving live data when there
are no properties data stored (hopefully only very old stacks) or when
fields other than properties (e.g. metadata) contain intrinsic functions.

Change-Id: Ib443cebc7b832a6cdb6583bb6cc830452357c392
Story: #2003987
Task: 26930