Don't assert that we haven't signalled the calling thread
We use the notify.signal() call to notify the main thread that the stack
status has moved to IN_PROGRESS so that it can wait before returning
control to the user. Therefore it is expected that if an operation
eventually fails (or succeeds), signal() will have been called a long
time previously. The only reason it is there is to guard against
failures before the resource attains the IN_PROGRESS state, where the
persistence of the state has been deferred to coincide with the lock
release.

Story #2003988
Task: 26931

Change-Id: Ie0519ee78607f71855c2c0ace2cb4ff52c5809b6