Use resource description as default description property
We allow users to provide a description for a resource that shows up in
the Heat API. Many resource types also allow users to add descriptions,
and these have typically been exposed as properties rather than by
trying to read the resource's own description.

To prevent the need to duplicate information, use the resource's
description as the default for any top-level properties named
'description'.

One downside of this is that any changes to the resource description
could cause updates to the resource. (However, there are no issues
specific to updgrades, because the subsitution is also done on the
previous properties.) To guard against the worst of this, only enable
this behaviour if the description can be updated without resource
replacement.

Change-Id: I56560d014a02b5f2ddbc08689d39147fbe4ffca4