Allow users to disable the tests run by Grenade
Setting RUN_HEAT_INTEGRATION_TESTS=False disables the Heat
integration tests which are normally executed by Grenade
after the deployment of the base environment and after the upgrade.
This is useful when Heat is used in a Grenade job of another
component, where the focus of the testing is not Heat itself,
thus shortening the run-time of the overall Grenade job.

The default behavior is unchanged.

Change-Id: I47b258fecd45ebc08c82df179625bcfb57a32894